Bumped to 3.0.0
Major overhaul. Numerous subtle bugs were fixed, most/all of which were runtime and some of which could easily reach production unnoticed, resulting in e.g. parts of the application silently disappearing, features being shut off, development logs becoming visible to end-users, and/or the delivery of payloads that exclude browsers the application is required to support (due to task runners/transpilers overlooking environment variables that indicate that the build should include additional polyfills and/or extra transpilation steps for supporting less modern browsers). Thus, any usages that relied on broken exported values stop working as expected. It's unlikely that many consumers of the library ran into this issue - almost all of the issues related to the less convenient (albeit usually more accurate) aliases, so it's more likely that these features would simply not be used (they are rarely used in the internal production codebases for which this was developed). I came across only one in our many codebases - which uses exports from this library 1000s of times - and discovered it before it reached production (this is what prompted the audit and repair of this library - the issue was very difficult to detect, and it caused an otherwise inexplicable show-stopping bug that took an extremely long time to trace to env-var-helpers). There's thus at least some potential for breakage since it's conceivable that someone may have run into our my issue but worked around it, so I felt a full major version upgrade was called for.